pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hashguard"
version = "1.1.2"
description = "A powerful, professional file verification and threat intelligence platform with VirusTotal integration."
requires-python = ">=3.9"
authors = [
    { name = "Alberto Tijunelis Neto", email = "albertotijunelis@gmail.com" }
]
readme = "README.md"
license = { text = "Elastic-2.0" }
keywords = ["malware", "security", "verification", "threat-intelligence", "virustotal", "file-analysis", "pe-analysis", "yara", "malwarebazaar"]
urls = { Homepage = "https://github.com/albertotijunelis/hashguard", Repository = "https://github.com/albertotijunelis/hashguard", Issues = "https://github.com/albertotijunelis/hashguard/issues" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: Utilities",
]

dependencies = [
    "requests>=2.28.0",
    "pefile>=2023.2.7",
    "yara-python>=4.3.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.22.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "aiofiles>=23.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "ppdeep>=20200403",
    "psutil>=5.9.0",
    "pyzipper>=0.3.6",
]

[project.scripts]
hashguard = "hashguard.cli:main"
hashguard-web = "hashguard.web.api:start_server"

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-cov", "httpx"]
build = ["pyinstaller>=5.0"]
dev = ["black", "pylint", "mypy"]
full = ["lief>=0.13.0", "tlsh>=4.5.0", "networkx>=3.0", "stix2>=3.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
hashguard = [
    "data/*.json",
    "yara_rules/*.yar",
    "yara_rules/*.yara",
    "assets/*.png",
    "assets/*.ico",
    "web/templates/*.html",
    "web/static/*.png",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=hashguard --cov-report=term-missing --cov-report=xml"
filterwarnings = [
    "ignore::requests.exceptions.RequestsDependencyWarning",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
